requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.0
easyocr==1.7.1
filelock==3.12.4
fonttools==4.43.0
idna==3.4
imageio==2.31.5
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.26.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
pyclipper==1.3.0.post5
pyparsing==3.1.1
pypylon==2.3.0
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.3
shapely==2.0.1
six==1.16.0
sympy==1.12
tifffile==2023.9.26
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.8.0
urllib3==2.0.6
